Rollbar collector first draft
For now this is just a count of item occurrences in a single rollbar
project, across all environments. More fine grained control is needed.
Ideally you would give it a list of access tokens, in order to get stats
from multiple projects, and it would be nice to use the project name
somewhere in the stat (or as a tag).

Also this needs some sort of sanity check to prevent endless fetches if
for some reason rollbar doesn't send an item back with a certain id.
Maybe a better check would be greater than/less than, rather than
equals.